# server-tv-pairing/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flint device pairing API",
    "version": "1.1.0",
    "description": "Encrypted VPN configuration handoff for Android, Android TV, Windows and iOS. No account session is transferred. Owner QR: HTTPS /connect/{id}#key=...; receiver QR: flint://pair?v=1&id=...#key=... . The production server has not been changed by this package."
  },
  "servers": [
    {
      "url": "https://flintmain.ru"
    }
  ],
  "paths": {
...
    "/api/v1/devices/pairing/share/start": {
      "post": {
        "operationId": "DevicePairingShareStart",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareStart"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid login or TV proof",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Subscription is not owned or not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "QR already approved or not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "QR expired, cancelled or acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareStarted"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/pairing/share/claim": {
      "post": {
        "operationId": "DevicePairingShareClaim",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareClaim"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid login or TV proof",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Subscription is not owned or not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "QR already approved or not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "QR expired, cancelled or acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Completed"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/pairing/share/cancel": {
      "post": {
        "operationId": "DevicePairingShareCancel",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inspect"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid login or TV proof",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Subscription is not owned or not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "QR already approved or not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "QR expired, cancelled or acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Device": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "type": "string",
            "maxLength": 128
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "android-tv",
              "windows",
              "ios"
            ]
          },
          "model": {
            "type": "string",
            "maxLength": 120
          },
          "osVersion": {
            "type": "string",
            "maxLength": 40
          },
          "appVersion": {
            "type": "string",
            "maxLength": 32
          }
        },
        "required": [
          "deviceId",
          "platform",
          "model"
        ]
      },
      "Envelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "nonce": {
            "type": "string",
            "description": "base64url, no padding; 12 bytes"
          },
          "ciphertext": {
            "type": "string",
            "maxLength": 699076,
            "description": "base64url AES-256-GCM ciphertext including 16-byte tag"
          }
        },
        "required": [
          "version",
          "nonce",
          "ciphertext"
        ]
      },
...
      "Inspect": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pairingId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          }
        },
        "required": [
          "pairingId"
        ]
      },
...
      "Completed": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "encryptedSettings": {
            "$ref": "#/components/schemas/Envelope"
          }
        },
        "required": [
          "encryptedSettings"
        ]
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "ShareStart": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "subscriptionId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "claimChallenge": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          }
        },
        "required": [
          "subscriptionId",
          "requestId",
          "claimChallenge"
        ]
      },
      "ShareStarted": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pairingId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresInSeconds": {
            "type": "integer",
            "maximum": 900
          }
        },
        "required": [
          "pairingId",
          "expiresAt",
          "expiresInSeconds"
        ]
      },
      "ShareClaim": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pairingId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "claimSecret": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "codeChallenge": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "device": {
            "$ref": "#/components/schemas/Device"
          }
        },
        "required": [
          "pairingId",
          "claimSecret",
          "codeChallenge",
          "device"
        ]
      }
    }
  }
}